Add files via upload
This is a copy of my code to solve the Mountain car v0 problem with discrete action. I run it on a standalone machine, with no execution problem. However, the main problem is every time I run this same code, I get different results (i.e., different average scores). When I render it in human mode and test the trained model, sometimes I see the agent reach the flag, and sometimes not. I have tried to change the reward shape, the epsilon values, the number of episodes, the buffer limit, and so on, and yet this problem I mentioned above happens. What is the solution to this issue?